final_pjt/Finance_Service/final-pjt-back/requirements.txt
# appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asgiref==3.7.2
# asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==23.1.0
# backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
beautifulsoup4==4.12.2
# Bottleneck @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_29949159-f86f-474b-bc1f-aaa1e0e222b4ofusifik/croots/recipe/bottleneck_1657175564045/work
# Brotli @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_01vdzesc6m/croots/recipe/brotli-split_1659616063272/work
bs4==0.0.1
# certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2emnmf1qow/croot/certifi_1690232232440/work/certifi
cffi==1.16.0
charset-normalizer==3.3.2
# comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
# debugpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_30hp2nowkm/croot/debugpy_1690905056188/work
# decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
# exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
# executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fonttools==4.44.3
h11==0.14.0
# idna @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_00jf0h4zbt/croot/idna_1666125573348/work
# importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.1.1
# ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1698244280508/work
# ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1698846678920/work
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
# jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
# jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1698673726529/work
kiwisolver==1.4.5
matplotlib==3.8.1
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
# mkl-fft @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_af8wl3mjfu/croot/mkl_fft_1695058185839/work
# mkl-random @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_0fq2s894lk/croot/mkl_random_1695059822529/work
mkl-service==2.4.0
# nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
# numexpr @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_c2pokmr868/croot/numexpr_1696515300257/work
numpy==1.26.2
oauthlib==3.2.2
outcome==1.3.0.post0
# packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.3
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
# pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.1.0
# platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1699715570510/work
# prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1699963054032/work
# psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1695367144488/work
# ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
# pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
PyJWT==2.8.0
# pyOpenSSL @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_7fwdg30481/croot/pyopenssl_1690223429070/work
pyparsing==3.1.1
# PySocks @ file:///opt/concourse/worker/volumes/live/112288ac-9cb0-4e73-768b-13baf4ca6419/volume/pysocks_1605305820043/work
# python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python3-openid==3.2.0
pytz==2023.3.post1
# pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1666828678061/work
# requests @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_fa2zrx4csg/croot/requests_1690400223595/work
requests-oauthlib==1.3.1
selenium==4.15.2
# six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
# stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
# tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1695373639474/work
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698671135544/work
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
# tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==2.1.0
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1699959196938/work
wsproto==1.2.0
# zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
